Project Description
This will be the main object for this project.  Project will allow you
to host a JS file. The JS file will have a self created token.  This is
to better Stay connected to the same results.  This JS file has one
job.  To see everything that is on the page, create an object and read
where the viewer is on the page and how long they view it.  Where they
clicked, and where they went when they left.  The idea here is to have
a better understanding of where our viewers go and what they look at.